gst/msdk: move JPEG encoder test classes to lib
